+++ app/components/github/example_component.html.erb
