+++ app/components/github/repo_tree_component.html.erb
@@ -4,9 +4,11 @@
     <% if item["type"] == "file" %>
        <li><%= link_to item["name"], item["html_url"], class: "hover:bg-gray-700 block p-2" %></li>
     <% else %>
+    <li>
       <span class="inline-flex justify-center items-center ml-4">
         <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M3 12l2-2m0 0l7-7 7 7M5 10v10a1 1 0 001 1h3m10-11l2 2m-2-2v10a1 1 0 01-1 1h-3m-6 0a1 1 0 001-1v-4a1 1 0 011-1h2a1 1 0 011 1v4a1 1 0 001 1m-6 0h6"></path></svg>
       </span>
-      <%= link_to "v #{item['name']}", pages_repo_path(github_repo: @git_repo, path: path_content(item["url"])) %>
+      <%= link_to "#{item['name']}", pages_repo_path(github_repo: @git_repo, path: path_content(item["url"])) %>
+    </li>
     <% end %>
 <% end %>
