+++ app/components/github/url_search_component.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+module Github
+class UrlSearchComponent < ViewComponent::Base
+
+end
+end
