--- app/components/github_url_search_component.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class GithubUrlSearchComponent < ViewComponent::Base
-
-end
