+++ app/views/pages/home.html.erb
@@ -1,1 +1,1 @@
-<%= render GithubUrlSearchComponent.new %>
+<%= render Github::UrlSearchComponent.new %>
